requirements.txt
# 基础依赖
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.26.0

# 数据库
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Web框架
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# 数据处理和分析
scikit-learn>=0.24.0
optuna>=2.10.0
xgboost>=1.5.0
lightgbm>=3.3.0

# 可视化
plotly>=5.3.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# 交易相关
ib_insync>=0.9.70
ta-lib>=0.4.24

# 异步支持
aiohttp>=3.8.0
asyncio>=3.4.3

# 工具库
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5

# 其他依赖
python-telegram-bot>=20.0
streamlit>=1.0.0
yfinance>=0.1.63
ta>=0.7.0
python-dateutil>=2.8.2
pytz>=2021.1
statsmodels>=0.13.0
scipy>=1.7.0
tensorflow>=2.6.0
keras>=2.6.0
bcrypt>=3.2.0
smtplib3>=0.0.1
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910 